CASSANDRA-20259: Refactor to decouple RowIterator and CellIterator from Spark so bulk reads can be performed outside of Spark
