+++ cassandra-analytics-core/src/main/spark2/org/apache/cassandra/spark/sparksql/SparkRowIterator.java
@@ -21,6 +21,7 @@ package org.apache.cassandra.spark.sparksql;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.function.Function;
 
 import com.google.common.annotations.VisibleForTesting;
 
@@ -28,6 +29,7 @@ import org.apache.cassandra.spark.config.SchemaFeature;
 import org.apache.cassandra.spark.data.DataLayer;
 import org.apache.cassandra.spark.sparksql.filters.PartitionKeyFilter;
 import org.apache.spark.sql.catalyst.InternalRow;
+import org.apache.spark.sql.catalyst.expressions.GenericInternalRow;
 import org.apache.spark.sql.sources.v2.reader.InputPartitionReader;
 import org.apache.spark.sql.types.StructType;
 import org.jetbrains.annotations.NotNull;
@@ -36,7 +38,7 @@ import org.jetbrains.annotations.Nullable;
 /**
  * Wrapper iterator around SparkCellIterator to normalize cells into Spark SQL rows
  */
-public class SparkRowIterator extends AbstractSparkRowIterator implements InputPartitionReader<InternalRow>
+public class SparkRowIterator extends AbstractSparkRowIterator<InternalRow> implements InputPartitionReader<InternalRow>
 {
     @VisibleForTesting
     public SparkRowIterator(int partitionId, @NotNull DataLayer dataLayer)
@@ -49,19 +51,37 @@ public class SparkRowIterator extends AbstractSparkRowIterator implements InputP
                                @Nullable StructType requiredSchema,
                                @NotNull List<PartitionKeyFilter> partitionKeyFilters)
     {
-        super(partitionId, dataLayer, requiredSchema, partitionKeyFilters);
+        super(partitionId,
+              dataLayer,
+              requiredSchema,
+              partitionKeyFilters,
+              (builder) -> decorate(builder, dataLayer.requestedFeatures()));
     }
 
     @Override
     @NotNull
-    RowBuilder newBuilder()
+    protected RowBuilder<InternalRow> newBuilder(Function<RowBuilder<InternalRow>, RowBuilder<InternalRow>> decorator)
     {
-        RowBuilder builder = new FullRowBuilder(cqlTable, hasProjectedValueColumns);
-        for (SchemaFeature feature : requestedFeatures)
+        RowBuilder<InternalRow> builder = new FullRowBuilder<>(cqlTable, hasProjectedValueColumns, this::mapper);
+        builder = decorator.apply(builder);
+        builder.reset();
+        return builder;
+    }
+
+    protected static RowBuilder<InternalRow> decorate(RowBuilder<InternalRow> builder,
+                                                      List<SchemaFeature> features)
+    {
+        for (SchemaFeature feature : features)
         {
             builder = feature.decorate(builder);
         }
-        builder.reset();
+
         return builder;
+    }
+
+    @Override
+    public InternalRow mapper(Object[] result)
+    {
+        return new GenericInternalRow((result));
     }
 }
